Enable debug at local
Add Xcode project which make the local debugging possible. To debug at local, the LOCAL_MACHINE in edmodes.h should set as 1. Besides, for local debugging, the config files with _local_ will be called rather than the original one. To run it on cluster, type in the command, which is saved in edmodels.h, in terminal, then it will be copied to cluster, and you need to type 'make' in cluster terminal to run it.