Version archive
This version is used to estimate 2007-2016 carbon fluxes for NASA IDS project at Jan 16th 2020, corresponding to results (test_glb_newPFT_soilgrids_v15*) saved on department clusters. This version did have spun-up from bare ground and also the run with land-use change. This version also add additional code to calculate RH metrics, which may be remove later if they are not helpful for global initialization. This commit is for archiving a version for Lesley project.